# docs/my-bike-log-swagger.yml
openapi: 3.0.3
info:
  title: BikeLog API
  description: >
    REST API contract for the BikeLog application (migrated from a static
    localStorage-based HTML app to a multi-user Spring Boot + MySQL backend).
    Supports Google OAuth 2.0 login, JWT-secured access, multiple bikes per
    user, ride/fuel/maintenance tracking, and pre-computed dashboard statistics.
  version: 1.0.0
servers:
  - url: https://api.bikelog.example.com/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /rides/{bikeId}:
    parameters:
      - $ref: '#/components/parameters/BikeId'
    get:
      tags: [Rides]
      summary: List rides for a bike
      parameters:
        - name: month
          in: query
          description: Filter to a specific month, format YYYY-MM.
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Paginated list of rides, newest first. Each ride includes
            the computed distance covered since the previous odometer reading.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RidePage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Rides]
      summary: Log a new ride
      description: >
        `odo` must be greater than or equal to the bike's current total
        odometer (max of initialOdo, all ride odo, and all petrol-entry odo
        values). Violations return 409 Conflict.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RideCreateRequest'
      responses:
        '201':
          description: Ride created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ride'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Odometer reading is less than the bike's current total odometer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rides/{bikeId}/{rideId}:
    parameters:
      - $ref: '#/components/parameters/BikeId'
      - name: rideId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [Rides]
      summary: Delete a ride
      responses:
        '204':
          description: Ride deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ───────────────────────── PETROL ─────────────────────────
  /petrol-entries/{bikeId}:
    parameters:
      - $ref: '#/components/parameters/BikeId'
    get:
      tags: [Petrol]
      summary: List petrol fill-up entries for a bike
      parameters:
        - name: month
          in: query
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: >
            Paginated list of petrol entries, newest first. Each entry
            includes the computed `litres` (amount / pricePerLitre), 
            `distanceKm` represents the distance travelled using the fuel from 
            this petrol entry. the running `cumulativeLitres` for the bike, and 
            `mileageKmPerLitre` represents the mileage achieved using the fuel 
            from this petrol entry and is calculated as distanceKm / litres.
            The latest petrol entry will have null values for distanceKm and 
            mileageKmPerLitre until the next petrol refill is logged.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetrolEntryPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Petrol]
      summary: Log a new petrol fill-up
      description: >
        `litres` is computed server-side as `amount / pricePerLitre` and
        rounded to 2 decimal places. `odo` must be greater than or equal to
        the bike's current total odometer. After a new petrol entry is created,
        the previous petrol entry is updated with distanceKm and 
        mileageKmPerLitre using the difference between the two odometer readings.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PetrolEntryCreateRequest'
      responses:
        '201':
          description: Petrol entry created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetrolEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Odometer reading is less than the bike's current total odometer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /petrol-entries/{bikeId}/{petrolEntryId}:
    parameters:
      - $ref: '#/components/parameters/BikeId'
      - name: petrolEntryId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [Petrol]
      summary: Delete a petrol entry
      responses:
        '204':
          description: Petrol entry deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ───────────────────────── OIL CHANGES ─────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT issued after successful Google OAuth 2.0 login. Send as
        `Authorization: Bearer <token>` on every request except the
        /auth/google/* endpoints.
  parameters:
    BikeId:
      name: bikeId
      in: path
      required: true
      description: ID of the bike, must be owned by the authenticated user.
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      description: Zero-based page number.
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      description: Page size.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
  responses:
    BadRequest:
      description: Validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    Unauthorized:
      description: Missing, invalid, or expired JWT.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found, or not owned by the current user.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 404
        error:
          type: string
          example: Not Found
        message:
          type: string
          example: Bike not found or does not belong to the current user.
        path:
          type: string
          example: /api/v1/bikes/9b1f.../rides
    ValidationFieldError:
      type: object
      properties:
        field:
          type: string
          example: odo
        message:
          type: string
          example: Odometer must be greater than or equal to 12540.00 km
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            fieldErrors:
              type: array
              items:
                $ref: '#/components/schemas/ValidationFieldError'

    # ── Auth / Users ──
...
    Ride:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bikeId:
          type: string
          format: uuid
        date:
          type: string
          format: date
        time:
          type: string
          example: "18:45"
        odo:
          type: number
          format: double
        distanceKm:
          type: number
          format: double
          description: Computed distance since the previous odometer reading (>= 0).
        createdAt:
          type: string
          format: date-time
      required: [id, bikeId, date, time, odo, distanceKm]
    RideCreateRequest:
      type: object
      required: [date, time, odo]
      properties:
        date:
          type: string
          format: date
        time:
          type: string
          pattern: '^([01]\d|2[0-3]):[0-5]\d$'
          example: "18:45"
        odo:
          type: number
          format: double
          exclusiveMinimum: true
          minimum: 0
    RidePage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Ride'

    # ── Petrol ──
    PetrolEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bikeId:
          type: string
          format: uuid
        date:
          type: string
          format: date
        odo:
          type: number
          format: double
        amount:
          type: number
          format: double
          description: Amount paid, in ₹.
        pricePerLitre:
          type: number
          format: double
        litres:
          type: number
          format: double
          description: Computed as amount / pricePerLitre, rounded to 2 decimals.
        cumulativeLitres:
          type: number
          format: double
          description: Running total of litres for the bike, up to and including this entry.
        mileageKmPerLitre:
          type: number
          format: double
          nullable: true
          description: >
            Mileage achieved using the fuel from this petrol entry.
            Calculated as distanceKm / litres.
            Will be null for the latest petrol entry.
        distanceKm:
          type: number
          format: double
          nullable: true
          description: >
            Distance travelled using the fuel from this petrol entry.
            will be null for the latest petrol entry.
        createdAt:
          type: string
          format: date-time
      required: [id, bikeId, date, odo, amount, pricePerLitre, litres, cumulativeLitres]
    PetrolEntryCreateRequest:
      type: object
      required: [date, odo, amount, pricePerLitre]
      properties:
        date:
          type: string
          format: date
        odo:
          type: number
          format: double
          exclusiveMinimum: true
          minimum: 0
        amount:
          type: number
          format: double
          exclusiveMinimum: true
          minimum: 0
        pricePerLitre:
          type: number
          format: double
          exclusiveMinimum: true
          minimum: 0
          example: 115.81
    PetrolEntryPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/PetrolEntry'

    # ── Oil Changes ──
...
    PageMeta:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer

    # ── Dashboard / Stats ──
...